size_t for matlab band.cpp also works with R

+++ gpds_cpp/band.cpp
@@ -1,6 +1,6 @@
 #include "band.h"
 
-typedef int blasint;
+typedef size_t blasint;
 
 extern "C" {
